fix test and fail when test fails

+++ test/run_tests.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env sh
 BASEDIR=$(dirname "$0")
-cd "$BASEDIR" || exit
+cd "$BASEDIR" || exit 1
+
+RED='\033[0;31m'
+GREEN='\033[0;32m'
 
 
 # make a temp dir and copy the input file to it
@@ -11,6 +14,11 @@ cp ./input.vm "$TEMPDIR"/input.vm
 yes | ../report-updater.sh "$TEMPDIR"
 
 # check if the output matches what we expected
-diff --side-by-side "$TEMPDIR"/input.vm ./output.vm
-
-rm -rf "$TEMPDIR"
+if diff --side-by-side "$TEMPDIR"/input.vm ./output.vm; then
+    rm -rf "$TEMPDIR"
+    echo $GREEN"passed"
+else
+    #rm -rf "$TEMPDIR"
+    echo $RED"failed"
+    exit 1
+fi
